+++ CHANGELOG.md
@@ -12,6 +12,7 @@ All notable changes to this project will be documented in this file.
 - Bugfix and improvements
 - Transponder frequency and symbol-rate precision
 - Channel list format and conversion, better visual feedback
+- CLI improvements: tools, macro
 - Updated translations